+++ src/main/java/com/sim/wicmsapi/service/ContentTypeService.java
@@ -7,7 +7,7 @@ import com.sim.wicmsapi.entity.ContentType;
 
 public interface ContentTypeService {
 
-	public Optional<ContentType> getContentType(long contentId);
+	public Optional<ContentType> getContentType(int contentId);
 	public List<ContentType> getCTs();
 	public ContentType save(ContentType contentType);
 	public List<ContentType> getCTs(long ctTypeId,String active);
